clean up Maker interface
